model-builder/requirements.txt
implicit==0.7.2
pandas==1.3.5
python-dotenv==0.21.1
s3fs==0.4.2
numpy==1.21.6
fastapi==0.103.2
uvicorn==0.22.0
lightfm==1.17
boto3==1.33.13
scipy==1.7.3
